Add `_post_coinit/activeobj.py` and move `GetActiveObject` from `misc.py`.
